move setup to integration spec
