+++ spec/fakes/bin/hub
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+
+require File.expand_path(File.join('..', '..', 'support', 'fake_github'), File.dirname(__FILE__))
+
+FakeGithub.new(ARGV).run!
+binding.pry
