+++ spec/support/aruba.rb
@@ -1,0 +1,3 @@
+ENV ||= {}
+ENV['TESTING'] = 'true'
+@aruba_timeout_seconds = 560
